+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - unified custom features
+- money management for cosmere/nimble characters
 
 ## [0.5.10] - 2026-09-07
 ### Added